Simulation: recursive resolver implementation with channels
